# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: VertexLearn AI API
  version: 0.7.0
  description: Phase 7 surface — auth, course foundation, learning core, AI tutor/RAG (course-isolated), summaries, AI quiz drafts, flashcards, study plans, mastery, recommendations, admin workspace, course approvals, discussions, announcements, notifications, email abstraction, S3/MinIO storage (uploads, signed URLs), catalog filters/reviews, payments/revenue, Redis caching (catalog/course/leaderboard, x-cache), rate limits (auth 10/min/IP, AI 20/min/user, global 100/min), gamification leaderboard, health/ready aliases.
servers:
  - url: http://localhost:4000
paths:
...
  /api/v1/courses:
    get:
      summary: Public catalog (paginated, searchable, filterable by status/category/difficulty/minRating; includes avg_rating/rating_count; Redis-cached 60s with x-cache HIT/MISS)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20 } }
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [draft, pending, published, rejected] } }
        - { name: category, in: query, schema: { type: string } }
        - { name: difficulty, in: query, schema: { type: string, enum: [beginner, intermediate, advanced] } }
        - { name: minRating, in: query, schema: { type: number, minimum: 0, maximum: 5 } }
      responses:
        '200': { description: Paginated courses with rating aggregates }
        '429': { description: Rate limited (global 100/min backstop) }
    post:
      summary: Create course (instructor|admin, status defaults to pending)
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
  /api/v1/courses/{id}:
    get:
      summary: Public course detail with modules + lectures
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Course detail }
        '404': { description: Not found }
    put:
      summary: Update own course (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated }
        '403': { description: Not the owner }
        '404': { description: Not found }
  /api/v1/courses/{id}/modules:
    post:
      summary: Add module to own course
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Created }
        '403': { description: Not the owner }
  /api/v1/courses/modules/{id}/lectures:
    post:
      summary: Add lecture to own module
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Created }
        '403': { description: Not the owner }
...
  /api/v1/courses/{id}/reviews:
    get:
      summary: Public review list with real aggregate (avg_rating, rating_count)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Reviews + aggregate }
        '404': { description: Course not found }
    post:
      summary: Submit/update my review (enrolled student, owner, or admin; one per student per course)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Review saved }
        '400': { description: Rating out of range }
        '403': { description: Enrollment required }
        '404': { description: Course not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT